+++ src/frequenz/channels/_anycast.py
@@ -12,7 +12,7 @@ from typing import Deque, Generic, Optional
 from ._base_classes import ChannelClosedError
 from ._base_classes import Receiver as BaseReceiver
 from ._base_classes import Sender as BaseSender
-from ._base_classes import T
+from ._base_classes import SenderError, T
 
 
 class Anycast(Generic[T]):
@@ -123,7 +123,7 @@ class Sender(BaseSender[T]):
         """
         self._chan = chan
 
-    async def send(self, msg: T) -> bool:
+    async def send(self, msg: T) -> None:
         """Send a message across the channel.
 
         To send, this method inserts the message into the Anycast channel's
@@ -134,19 +134,21 @@ class Sender(BaseSender[T]):
         Args:
             msg: The message to be sent.
 
-        Returns:
-            Whether the message was sent, based on whether the channel is open
-                or not.
+        Raises:
+            SenderError: if the underlying channel was closed.
+                A [ChannelClosedError][frequenz.channels.ChannelClosedError] is
+                set as the cause.
         """
         if self._chan.closed:
-            return False
+            raise SenderError("The channel was closed", self) from ChannelClosedError(
+                self._chan
+            )
         while len(self._chan.deque) == self._chan.deque.maxlen:
             async with self._chan.send_cv:
                 await self._chan.send_cv.wait()
         self._chan.deque.append(msg)
         async with self._chan.recv_cv:
             self._chan.recv_cv.notify(1)
-        return True
 
 
 class Receiver(BaseReceiver[T]):