+++ RELEASE_NOTES.md
@@ -6,12 +6,15 @@
 
 ## Upgrading
 
-<!-- Here goes notes on how to upgrade from previous versions, including if there are any depractions and what they should be replaced with -->
+* The `Sender.send()` method now `raise`s a `SenderError` instead of returning `False`. The `SenderError` will typically have a `ChannelClosedError` and the underlying reason as a chained exception.
 
 ## New Features
 
-* A new base exception `frequenz.channels.Error` was added. All exceptions from
-  this library inherit from this exception.
+* New exceptions were added:
+
+  * `Error`: A base exception from which all exceptions from this library inherit.
+
+  * `SendError`: Raised for errors when sending messages.
 
 ## Bug Fixes